+++ x/slashing/types/msg.go
@@ -3,17 +3,16 @@ package types
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-)
-
-// slashing message types
-const (
-	TypeMsgUnjail = "unjail"
+	"github.com/cosmos/cosmos-sdk/x/auth/migrations/legacytx"
 )
 
 // verify interface at compile time
 var (
 	_ sdk.Msg = &MsgUnjail{}
 	_ sdk.Msg = &MsgUpdateParams{}
+
+	_ legacytx.LegacyMsg = &MsgUnjail{}
+	_ legacytx.LegacyMsg = &MsgUpdateParams{}
 )
 
 // NewMsgUnjail creates a new MsgUnjail instance
@@ -24,12 +23,6 @@ func NewMsgUnjail(validatorAddr sdk.ValAddress) *MsgUnjail {
 		ValidatorAddr: validatorAddr.String(),
 	}
 }
-
-// Route implements the sdk.Msg interface.
-func (msg MsgUnjail) Route() string { return RouterKey }
-
-// Type implements the sdk.Msg interface.
-func (msg MsgUnjail) Type() string { return TypeMsgUnjail }
 
 // GetSigners returns the expected signers for MsgUnjail.
 func (msg MsgUnjail) GetSigners() []sdk.AccAddress {